shoex-backend/requirements.txt
Django==5.2.6
graphene-django==3.2.2
django-cors-headers==4.6.0
django-filter==24.3
graphene-file-upload==1.3.0
psycopg2-binary==2.9.10
dj-database-url==2.3.0
Pillow==11.1.0
djangorestframework==3.15.2